Move `core/convar/` to `tier1/` 2 (#928)
