+++ primedev/client/clientruihooks.cpp
@@ -1,4 +1,4 @@
-#include "core/convar/convar.h"
+#include "tier1/convar.h"
 
 ConVar* Cvar_rui_drawEnable;
