+++ primedev/plugins/pluginmanager.cpp
@@ -4,7 +4,7 @@
 #include <ranges>
 #include "plugins.h"
 #include "config/profile.h"
-#include "core/convar/concommand.h"
+#include "tier1/cmd.h"
 
 namespace fs = std::filesystem;
