+++ primedev/scripts/client/scriptservertoclientstringcommand.cpp
@@ -1,6 +1,6 @@
 #include "squirrel/squirrel.h"
-#include "core/convar/convar.h"
-#include "core/convar/concommand.h"
+#include "tier1/convar.h"
+#include "tier1/cmd.h"
 
 void ConCommand_ns_script_servertoclientstringcommand(const CCommand& arg)
 {
