+++ primedev/server/auth/serverauthentication.cpp
@@ -1,12 +1,12 @@
 #include "serverauthentication.h"
 #include "shared/exploit_fixes/ns_limits.h"
-#include "core/convar/cvar.h"
-#include "core/convar/convar.h"
+#include "tier1/cvar.h"
+#include "tier1/convar.h"
 #include "masterserver/masterserver.h"
 #include "server/serverpresence.h"
 #include "engine/hoststate.h"
 #include "bansystem.h"
-#include "core/convar/concommand.h"
+#include "tier1/cmd.h"
 #include "dedicated/dedicated.h"
 #include "config/profile.h"
 #include "core/tier0.h"
