+++ primedev/server/idlekick.cpp
@@ -1,4 +1,4 @@
-#include "core/convar/convar.h"
+#include "tier1/convar.h"
 #include "dedicated/dedicated.h"
 #include "silver-bun/module.h"
