+++ primedev/shared/exploit_fixes/exploitfixes.cpp
@@ -1,4 +1,4 @@
-#include "core/convar/cvar.h"
+#include "tier1/cvar.h"
 #include "ns_limits.h"
 #include "dedicated/dedicated.h"
 #include "core/tier0.h"
