+++ primedev/tier1/convar.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "core/math/color.h"
 #include "cvar.h"
-#include "concommand.h"
+#include "cmd.h"
 
 // taken directly from iconvar.h
